move ground truth NSRTs to separate file
