+++ src/sampler_learning.py
@@ -10,6 +10,7 @@ from predicators.src import utils
 from predicators.src.torch_models import MLPClassifier, NeuralGaussianRegressor
 from predicators.src.settings import CFG
 from predicators.src.envs import create_env
+from predicators.src.ground_truth_nsrts import get_gt_nsrts
 
 
 def learn_samplers(strips_ops: List[STRIPSOperator],
@@ -49,8 +50,6 @@ def _extract_oracle_samplers(
     env = create_env(CFG.env)
     # We don't need to match ground truth NSRTs with no continuous
     # parameters, so we filter them out.
-    # Note: This import is in here to avoid cyclic imports.
-    from predicators.src.approaches.oracle_approach import get_gt_nsrts  # pylint: disable=import-outside-toplevel
     gt_nsrts = {
         nsrt
         for nsrt in get_gt_nsrts(env.predicates, env.options)
