+++ tests/test_planning.py
@@ -3,7 +3,7 @@
 import pytest
 from gym.spaces import Box
 from predicators.src.approaches import OracleApproach
-from predicators.src.approaches.oracle_approach import get_gt_nsrts
+from predicators.src.ground_truth_nsrts import get_gt_nsrts
 from predicators.src.approaches import ApproachFailure, ApproachTimeout
 from predicators.src.envs import CoverEnv
 from predicators.src.planning import sesame_plan, task_plan, task_plan_grounding
